Separate LinkBackAnchor and Breadcrumbs
Apparently, babel updates broke jest for no reason.
