+++ src/components/atoms/_index.scss
@@ -3,6 +3,7 @@
 @import './card';
 @import './codeBlock';
 @import './expertise';
+@import './linkBackAnchor';
 @import './pageGraphic';
 @import './pageSubtitle';
 @import './pageTitle';
