+++ src/components/atoms/linkBackAnchor/_index.scss
@@ -1,22 +1,25 @@
-a.link-back.has-more {
-  display: inline-block;
-  margin-right: 0;
-  margin-bottom: 0;
-}
-
-// Second link in breadcrumbs
-a.link-back-more {
+// Return to previous page link
+a.link-back {
   transition: 0.3s ease all;
   font-size: 1.125rem;
   line-height: 1.35;
-  display: inline-block;
-  margin: 0.875rem 1.25rem 0 0;
+  display: block;
+  margin: 0.875rem 1.25rem 1.5rem;
+  padding-left: 24px;
   position: relative;
   &, &:link, &:visited {
     color: var(--fore-color-dark);
   }
 
+  &:before {
+    position: absolute;
+    left: 0;
+    font-size: 20px;
+    line-height: 24px;
+  }
+
   @media screen and (max-width: calc(#{$layout-medium-breakpoint} - 1px)) {
+    margin-left: 0.875rem;
     margin-right: 0.875rem;
   }
 }
